Now gets far enough in test rom to be able to print out name of rom and load tiles

+++ src/ui/io/graphics.rs
@@ -2,7 +2,7 @@ use crate::hardware::io::ppu::Colour;
 use crate::hardware::io::LcdPixels;
 use crate::hardware::io::{HEIGHT, WIDTH};
 
-pub const PIXEL_SIZE: u32 = 2;
+pub const PIXEL_SIZE: usize = 2;
 
 pub fn convert_gameboy_to_rgb565(gameboy: LcdPixels) -> Vec<u8> {
     let mut result: Vec<u16> = vec![];
@@ -18,3 +18,27 @@ pub fn convert_gameboy_to_rgb565(gameboy: LcdPixels) -> Vec<u8> {
 
     result.iter().flat_map(|&x| x.to_ne_bytes().to_vec()).collect()
 }
+
+// pub fn convert_gameboy_to_rgb565(gameboy: LcdPixels) -> [u8; WIDTH * HEIGHT * PIXEL_SIZE] {
+//     let mut result = [0; WIDTH * HEIGHT * PIXEL_SIZE];
+
+//     for i in 0..gameboy.len() {
+//         let pixel = gameboy[i];
+
+//         result[i] = match pixel {
+//             Colour::White => 0xFF,
+//             Colour::LightGrey => 0x73,
+//             Colour::DarkGrey => 0x4B,
+//             Colour::Black => 0x00,
+//         };
+
+//         result[i + 1] = match pixel {
+//             Colour::White => 0xFF,
+//             Colour::LightGrey => 0xB5,
+//             Colour::DarkGrey => 0x6B,
+//             Colour::Black => 0x00,
+//         };
+//     }
+
+//     result
+// }
